add: patron buttons. format: stylus finally. axios'ed the supporters list for easy updating.
