+++ docs/.vuepress/theme/styles/download.styl
@@ -1,8 +1,8 @@
-.download-tables
+.download-tables {
   max-width: 800px;
   margin-top: 20px;
   position: relative;
-  
+
   h3 {
     font-size: 16px;
     line-height: 1em;
@@ -105,7 +105,7 @@
           flex: 0 0 100%;
           flex-wrap: wrap;
           margin-top: 10px;
-          
+
           .cta-btn {
             background-color: seagreen !important;
           }
@@ -124,3 +124,4 @@
       }
     }
   }
+}
