+++ docs/.vuepress/theme/styles/includes.styl
@@ -1,2 +1,2 @@
-@require 'palette'
-@require 'mixins'
+@require 'palette';
+@require 'mixins';
