+++ docs/.vuepress/theme/styles/navbar.styl
@@ -1,12 +1,16 @@
-.navbar
-  .nav-links
-    .nav-item,
-    .repo-link
+.navbar {
+  .nav-links {
+    .nav-item, .repo-link {
       margin-left: 10px !important;
+    }
+  }
 
-  .search-box
+  .search-box {
     margin-right: 0 !important;
-    
-    input
+
+    input {
       // Prevent zoom on mobile
       font-size: 16px !important;
+    }
+  }
+}
