+++ docs/.vuepress/theme/styles/palette.styl
@@ -1,8 +1,8 @@
-$accentColor = dodgerblue
-$textColor = #2c3e50
-$borderColor = #eaecef
-$codeBgColor = #282c34
-$sidebarWidth = 220px
-$tableCodeBgColor = #ffeeec
-$tableCodeTextColor = #9b2e2e
-$tableAltBgColor = #f6f8fa
+$accentColor = dodgerblue;
+$textColor = #2c3e50;
+$borderColor = #eaecef;
+$codeBgColor = #282c34;
+$sidebarWidth = 220px;
+$tableCodeBgColor = #ffeeec;
+$tableCodeTextColor = #9b2e2e;
+$tableAltBgColor = #f6f8fa;
