+++ docs/.vuepress/theme/styles/sidebar.styl
@@ -1,24 +1,12 @@
-.sidebar
-  & > .sidebar-links
+.sidebar {
+  & > .sidebar-links {
     padding-bottom: 100px !important;
 
-    .is-sub-group
-      .sidebar-heading
+    .is-sub-group {
+      .sidebar-heading {
         font-weight: bold;
-        font-size: 16px
-
-  // .active,
-  // .sidebar-heading:hover
-  //   .arrow
-  //     &.right
-  //       border-left-color: $accentColor !important
-
-  //     &.down
-  //       border-top-color: $accentColor !important
-
-  // .arrow
-  //   &.right
-  //     border-left-color: $textColor !important
-
-  //   &.down
-  //     border-top-color: $textColor !important
+        font-size: 16px;
+      }
+    }
+  }
+}
