+++ docs/.vuepress/theme/styles/toc.styl
@@ -1,3 +1,5 @@
-.table-of-contents
-  .badge
-    vertical-align middle
+.table-of-contents {
+  .badge {
+    vertical-align: middle;
+  }
+}
