+++ docs/.vuepress/theme/styles/wrapper.styl
@@ -1,5 +1,5 @@
-$wrapper
-  max-width $contentWidth
-  margin 0 auto
-  padding 2rem 25px
-
+$wrapper {
+  max-width: $contentWidth;
+  margin: 0 auto;
+  padding: 2rem 25px;
+}
