+++ docs/README.md
@@ -22,8 +22,8 @@ MTGJSON is an open‐source project that catalogues all Magic cards in a portabl
 
 ::: tip Zach - Lead Developer &amp; Maintainer
 ![avatar](/images/avatar-zach.jpg "Zach")
-Zach is a 22‐year old hippo fanatic who loves playing with big data and giving back to the community in unique ways. He has worked on MTGJSON since 2016, and has led the design and development of version 4. His work can also be seen heavily in the open‐source Cockatrice game client.<br/>
-<i>All inquires about MTGJSON can be sent to <font color="blue">zach@mtgjson.com</font></i>
+Zach is a 22‐year old hippo fanatic who loves playing with big data and giving back to the community in unique ways. He has worked on MTGJSON since 2016, and has led the design and development of version 4. His work can also be seen heavily in the open‐source Cockatrice game client.<br/><br/>
+<em>All inquires about MTGJSON can be sent to <a href="mailto:zach@mtgjson.com">zach@mtgjson.com</a></em>
 :::
 
 ::: tip Martin - Web Developer
@@ -38,6 +38,6 @@ Eric is a conceptualizer that loves open-source software and cervidae. When he's
 
 ## Contributing
 
-Others have devoted their time and effort in to this project. If you have suggestions for improvements, bug reports, or would just like to help address existing issues, we are always looking for help. You can contribute to the project through the main [MTGJSON repo](https://github.com/mtgjson/mtgjson) or the [MTGJSON website repo](https://github.com/mtgjson/mtgjson-website). If you would like to help in other ways please consider donating to the project via [PayPal](https://www.paypal.me/Zachhalpern).
+Others have devoted their time and effort in to this project. If you have suggestions for improvements, bug reports, or would just like to help address existing issues, we are always looking for help. You can contribute to the project through the main [MTGJSON repo](https://github.com/mtgjson/mtgjson) or the [MTGJSON website repo](https://github.com/mtgjson/mtgjson-website). If you would like to help in other ways please consider donating to the project via <a class="inline-btn paypal" href="https://www.paypal.me/Zachhalpern">PayPal</a> (one-time donation) or via <a class="inline-btn patreon" href="https://www.patreon.com/MTGJSON">Patreon</a> (reoccurring donations).
 
-<GenerateServices/>
+<GenerateSupporters/>
